+++ inst/modules/marginal.R
@@ -34,19 +34,10 @@ marginal <- function(paper) {
     report <- "No effects were described with terms related to 'marginally significant'."
   } else if (tl == "red") {
     report_text <- "You described effects with terms related to 'marginally significant'. If *p* values above 0.05 are interpreted as an effect, you inflate the alpha level, and increase the Type 1 error rate. If a *p* value is higher than the prespecified alpha level, it should be interpreted as a non-significant result."
+
     guidance <- c(
       "For metascientific articles demonstrating the rate at which non-significant p-values are interpreted as marginally significant, see:",
-      bibentry(
-        bibtype = "Article",
-        title = "The Prevalence of Marginally Significant Results in Psychology Over Time",
-        author = "Olsson-Collentine, A., van Assen, M. A. L. M., & Hartgerink, C. H. J.",
-        year = "2019",
-        journal = "Psychological Science",
-        volume = 30,
-        issue = 4,
-        pages = "576--586",
-        doi = "10.1177/0956797619830326"
-      ) |> format(),
+      format_ref(marg_ref),
       "For the list of terms used to identifify marginally significant results, see this [blog post by Matthew Hankins](https://web.archive.org/web/20251001114321/https://mchankins.wordpress.com/2013/04/21/still-not-significant-2/)."
     )
 
@@ -69,3 +60,14 @@ marginal <- function(paper) {
   )
 }
 
+marg_ref <- bibentry(
+  bibtype = "Article",
+  title = "The Prevalence of Marginally Significant Results in Psychology Over Time",
+  author = "Olsson-Collentine, A., van Assen, M. A. L. M., & Hartgerink, C. H. J.",
+  year = "2019",
+  journal = "Psychological Science",
+  volume = 30,
+  issue = 4,
+  pages = "576--586",
+  doi = "10.1177/0956797619830326"
+)